requirements.txt
discord.py
requests
huggingface_hub
python-dotenv
torch
torchvision
torchaudio
transformers
accelerate